etl/requirements.txt
kafka-python
pymongo
psycopg2
